+++ src/shared/snippets/city-card.jsx
@@ -11,14 +11,14 @@ const CityCard = ({ data }) => {
               <figure className="img-wrapper">
                 <source
                   srcSet={
-                    data.cover_image ? data.cover_image[0].url :
+                    data.cover_image ? data.cover_image[0]?.url :
                     "/images/placeholder.svg"
                   }
                   media="(min-width: 62.5em)"
                 />
                 <img
                   srcSet={
-                    data.cover_image ? data.cover_image[0].url :
+                    data.cover_image ? data.cover_image[0]?.url :
                     "/images/placeholder.svg"
                   }
                   alt={data.title}