+++ src/shared/snippets/event-card.jsx
@@ -5,18 +5,18 @@ import dayjs from "dayjs";
 
 const EventCard = ({data}) => {
   return (
-    <div className="card-wrapper event-item mb-3">
+    <div className="card-wrapper article-item mb-3">
       <div className="card card-img no-after">
         <Link href={`/eventi/${data.slug}`} aria-label={data.title}>
           <div className="img-responsive-wrapper">
             <div className="img-responsive">
               <figure className="img-wrapper">
                 <source
-                  srcSet={data.cover_image ? data.cover_image[0].url : '/images/placeholder.svg'}
+                  srcSet={data.cover_image ? data.cover_image[0]?.url : '/images/placeholder.svg'}
                   media="(min-width: 62.5em)"
                 />
                 <img
-                  srcSet={data.cover_image ? data.cover_image[0].url : '/images/placeholder.svg'}
+                  srcSet={data.cover_image ? data.cover_image[0]?.url : '/images/placeholder.svg'}
                   className="p-card__image"
                   alt={data.title}
                 />
@@ -28,18 +28,15 @@ const EventCard = ({data}) => {
 
       <div className="card-body">
       <div className="category-top m-0">
-      {data?.area && data?.area.length > 0 ? (
-          <Link className="category" href={`/sezioni/${data?.area[0].slug}`}>
-            {data?.area[0].title}
+      {data?.category && data?.category.length > 0 ? (
+          <Link className="category" href={`/sezioni/${data?.category[0].slug}`}>
+            {data?.category[0].title}
           </Link>
         ) : null}
             <span className="data">{data.postDate}</span>
           </div>
         
-        <Link href={`/eventi/${data.slug}`} >
-          <h5 className="card-title serif underline">{data.title}</h5>
-        </Link>
-
+          <h5 className="card-title serif underline"><Link href={`/eventi/${data.slug}`}>{data.title}</Link></h5>
         <p className="card-text d-md-block d-none">{data.excerpt}</p>
       </div>
     </div>